Game state management (#6)
Adds a simple crate for managing game state. The user needs to provide an enum describing the different possible game states, add a struct as a resource and call update on the resource somewhere. See main.rs for an implementation example.

* Add initial game state library

* Example scene transitions in main

Not sure this is most ergonomic, might
be easier to just store variables as the
Kataster implementation does.